pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "proof-auditor"
version = "0.1.0"
description = "AI-powered mathematical proof verification via sorry diagnosis"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Boyue Tang"}
]

dependencies = [
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
